# contracts/catalog/openapi.json
{
  "components": {
    "schemas": {
...
      "DeviceState": {
        "description": "Runtime state of a device.",
        "properties": {
          "extra": {
            "additionalProperties": true,
            "description": "Additional device-specific state fields",
            "title": "Extra",
            "type": "object"
          },
          "input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Active input port",
            "title": "Input"
          },
          "power": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "True = ON, False = OFF",
            "title": "Power"
          }
        },
        "title": "DeviceState",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ManualStep": {
        "description": "A manual instruction surfaced by a topology manual node (e.g. set the Dodocus\nRCA hub to the LD position) — load-bearing when the activated path crosses a\nmanual switch (audio path through the hub).",
        "properties": {
          "instruction": {
            "description": "Human instruction to perform",
            "title": "Instruction",
            "type": "string"
          },
          "node": {
            "description": "Topology node id surfacing this instruction",
            "title": "Node",
            "type": "string"
          }
        },
        "required": [
          "node",
          "instruction"
        ],
        "title": "ManualStep",
        "type": "object"
      },
...
      "ScenarioState": {
        "description": "Runtime state of a scenario.",
        "properties": {
          "devices": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DeviceState"
            },
            "description": "Current state of all devices in the scenario",
            "title": "Devices",
            "type": "object"
          },
          "manual_steps": {
            "description": "Manual notes from the most recent activation (e.g. 'set the Dodocus to LD'); single source of truth — same data was previously duplicated in the SSE event payload + ScenarioResponse; survives page reload via /scenario/state.",
            "items": {
              "$ref": "#/components/schemas/ManualStep"
            },
            "title": "Manual Steps",
            "type": "array"
          },
          "scenario_id": {
            "description": "ID of the active scenario",
            "title": "Scenario Id",
            "type": "string"
          }
        },
        "required": [
          "scenario_id"
        ],
        "title": "ScenarioState",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "A web service that manages MQTT devices with typed configurations",
    "title": "MQTT Web Service",
    "version": "0.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/devices/{device_id}/persisted_state": {
      "get": {
        "description": "Get the persisted state of a specific device from the state store.\n\nArgs:\n    device_id: The ID of the device to retrieve state for\n    \nReturns:\n    JSON: The persisted device state\n    \nRaises:\n    HTTPException: If state persistence is not available or no state found",
        "operationId": "get_device_persisted_state_devices__device_id__persisted_state_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Device Persisted State Devices  Device Id  Persisted State Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Device Persisted State",
        "tags": [
          "State"
        ]
      }
    },
    "/devices/{device_id}/state": {
      "get": {
        "description": "Get information about a specific device's current state.\n\nArgs:\n    device_id: The ID of the device to retrieve\n    \nReturns:\n    Device-specific state (e.g., LgTvState, AppleTVState, EmotivaXMC2State, etc.)\n    that extends BaseDeviceState with device-specific fields like volume, power,\n    input sources, playback information, and other operational data.\n    \nRaises:\n    HTTPException: If device is not found or an error occurs",
        "operationId": "get_device_state_devices__device_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Device State",
        "tags": [
          "State"
        ]
      }
    },
...
    "/scenario/state": {
      "get": {
        "description": "Get the current scenario state.\n\nRooms are the concurrency unit (one active scenario per room). With `room` given,\nreturns that room's active scenario state. Without it: returns the single active\nscenario if exactly one room is active; 404 if none; 409 if several rooms are\nactive (pass `room` to disambiguate).\n\nReturns:\n    ScenarioState: Current scenario state\n\nRaises:\n    HTTPException: If no active scenario or service not initialized",
        "operationId": "get_scenario_state_scenario_state_get",
        "parameters": [
          {
            "description": "Room to read the active scenario of",
            "in": "query",
            "name": "room",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Room to read the active scenario of",
              "title": "Room"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scenario State",
        "tags": [
          "State"
        ]
      }
    },
...
    "/scenario/{scenario_id}/state": {
      "get": {
        "description": "Get the state of a specific scenario by ID.\n\nReturns the scenario state including device states if the scenario is currently active,\nor a basic state structure with empty device states if the scenario is inactive.\n\nArgs:\n    scenario_id: The ID of the scenario to retrieve state for\n    \nReturns:\n    ScenarioState: The scenario state with device information\n    \nRaises:\n    HTTPException: If scenario is not found or service not initialized",
        "operationId": "get_specific_scenario_state_scenario__scenario_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Specific Scenario State",
        "tags": [
          "State"
        ]
      }
    },
...
  }
}